+++ webpack.config.js
@@ -39,6 +39,10 @@ const extensionConfig = {
                     },
                 ],
             },
+            {
+                test: /\.css$/,
+                use: ['style-loader', 'css-loader']
+            }
         ],
     },
     devtool: 'nosources-source-map',